Rely default ConsoleRenderer constructor
This change should allow the default constructor of the ConsoleRenderer to determine if colors should be requested or not based on presence of colorama as a dependency. Should work the same if present but allow this library to be used out of the box without adding it as a requirement.